fix warnings

+++ mindspore/core/ops/median.cc
@@ -64,7 +64,7 @@ abstract::TupleShapePtr MedianInferShape(const PrimitivePtr &primitive,
                                          const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(primitive);
   auto x_shape = CheckAndConvertUtils::ConvertShapePtrToShapeMap(input_args[0]->BuildShape())[kShape];
-  int64_t x_size = x_shape.size();
+  int64_t x_size = static_cast<int64_t>(x_shape.size());
   std::vector<int64_t> out;
   auto check_global_median = primitive->GetAttr(kGlobalMedian);
   MS_EXCEPTION_IF_NULL(check_global_median);
@@ -88,7 +88,7 @@ abstract::TupleShapePtr MedianInferShape(const PrimitivePtr &primitive,
           out.push_back(1);
         }
       } else {
-        out.push_back(x_shape[i]);
+        out.push_back(static_cast<int64_t>(x_shape[i]));
       }
     }
   }
